requirements.txt
blinker==1.7.0
boto3==1.34.11
botocore==1.34.11
click==8.1.7
colorama==0.4.6
flask==3.0.0
h11==0.14.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
numpy==1.24.4
opencv-python==4.8.1.78
packaging==23.2
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.10.0
six==1.16.0
typing-extensions==4.9.0
urllib3==1.26.18
uvicorn==0.25.0
werkzeug==3.0.1
zipp==3.17.0
